feat：WAF、网站监控增加日志定时清理

+++ agent/app/dto/response/file.go
@@ -24,7 +24,7 @@ type FileTree struct {
 }
 
 type DirSizeRes struct {
-	Size float64 `json:"size" validate:"required"`
+	Size int64 `json:"size" validate:"required"`
 }
 
 type FileProcessKeys struct {
